backend/requirements.txt
flask==3.1.0
flask-cors==5.0.0
flask-pymongo==2.3.0
python-dotenv==1.0.1
pymongo==4.10.1
pillow==11.1.0
pytesseract==0.3.14
openai==1.59.7
requests==2.32.3
